consolidate cli packages
