+++ src/packages/cli/src/create/constants.ts
@@ -1,4 +1,4 @@
-import { version } from '../package.json';
+import { version } from '../../package.json';
 
 export const GRAPHWEAVER_TARGET_VERSION = version;
 export const MIKRO_ORM_TARGET_VERSION = '5.4.2';
