+++ src/packages/cli/src/create/template.ts
