--- src/packages/create-graphweaver/.gitignore
@@ -1,4 +1,0 @@
-test
-bin
-node_modules
-*.log
