--- src/packages/create-graphweaver/.prettierrc
@@ -1,7 +1,0 @@
-{
-	"useTabs": true,
-	"printWidth": 100,
-	"semi": true,
-	"singleQuote": true,
-	"trailingComma": "es5"
-}
