--- src/packages/create-graphweaver/README.md
@@ -1,7 +1,0 @@
-# `@exogee/create-graphweaver`
-
-## Getting Started
-
-```shell
-$ npm init @exogee/graphweaver
-```
